Move all unit tests to unit folder
