+++ tests/unit/api/v1/test_templates.py
