+++ tests/unit/core/cdktf/stacks/test_aws_stacks.py
@@ -5,7 +5,7 @@ import pytest
 from src.app.core.cdktf.stacks.base_stack import AbstractBaseStack
 from src.app.enums.regions import OpenLabsRegion
 from src.app.schemas.template_range_schema import TemplateRangeSchema
-from tests.core.cdktf.config import modify_cidr, one_all_template
+from tests.unit.core.cdktf.config import modify_cidr, one_all_template
 
 
 @pytest.fixture(scope="module")
