+++ tests/unit/crud/test_crud_host_templates.py
