+++ tests/unit/crud/test_crud_subnet_templates.py
