+++ tests/unit/crud/test_crud_vpc_templates.py
